// src/Types/IWebSites.ts
export interface IWebSites {
  twitter: {
    image: string;
    title: string;
  };
  facebook: {
    image: string;
    title: string;
  };
  linkedIn: {
    image: string;
    title: string;
  };
  telegram: {
    image: string;
    title: string;
  };
  instagram: {
    image: string;
    title: string;
  };
}